Fix #872: Move guava dependency from util to sail-model
