+++ core/sail/model/pom.xml
@@ -19,6 +19,10 @@
 			<artifactId>rdf4j-sail-api</artifactId>
 			<version>${project.version}</version>
 		</dependency>
+		<dependency>
+			<groupId>com.google.guava</groupId>
+			<artifactId>guava</artifactId>
+		</dependency>
 	</dependencies>
         <build>
                 <plugins>
